Add files via upload
Install the Required Libraries: In the terminal, run the following command to install the PyQt6 library:
pip install PyQt6
uses PyQt6 to create a basic data entry interface where users can enter a first name and last name. When they click 'Submit,' the data is saved in a CSV file, either creating it or appending to an existing file.

(2) uses PyQt6 to create a simple GUI for BMI calculation. The user inputs weight and height, clicks the button to calculate BMI, and the result is displayed in the GUI.